mapped relationship between user and tasks

+++ src/tasks/tasks.service.ts
@@ -1,3 +1,4 @@
+import { User } from './../auth/user.entity';
 import { TaskRepository } from './task.repository';
 import { GetTaskFilterDto } from './dto/get-task-filter.dto';
 import { CreateTaskDto } from './dto/create-task.dto';
@@ -5,7 +6,6 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
-import { UpdateResult } from 'typeorm';
 
 @Injectable()
 export class TasksService {
@@ -14,8 +14,8 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(taskFilterDto: GetTaskFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(taskFilterDto);
+  async getTasks(taskFilterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(taskFilterDto, user);
   }
 
   // getAllTasks() {
@@ -38,20 +38,22 @@ export class TasksService {
   //   return tasks;
   // }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) {
       throw new NotFoundException(`Task with ID: "${id}" not found.`);
     }
     return found;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return await this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return await this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID: "${id}" not found.`);
     }
@@ -60,11 +62,12 @@ export class TasksService {
   async updateTaskStatus(
     id: number,
     status: TaskStatus,
-  ): Promise<UpdateResult> {
-    return await this.taskRepository.update(id, { status: status });
-    // const task = await this.getTaskById(id);
-    // task.status = status;
-    // await task.save();
-    // return task;
+    user: User,
+  ): Promise<Task> {
+    // return await this.taskRepository.update(id, { status, userId: user.id }); // Promise<UpdateResult>
+    const task = await this.getTaskById(id, user);
+    task.status = status;
+    await task.save();
+    return task;
   }
 }
